feat: 불필요한 mapRow 제거
1. update, insert는 mapRow 메서드가 불필요하다. findAll은 setValues 메서드가 불필요하다.
2. 불필요한 구현을 제거하기 위해 mapRow, setValues를 인터페이스로 분리해보자.
