+++ app/src/main/java/com/techcourse/dao/UserDao.java
@@ -1,9 +1,9 @@
 package com.techcourse.dao;
 
+import com.interface21.jdbc.core.JdbcTemplate;
+import com.interface21.jdbc.core.PreparedStatementSetter;
+import com.interface21.jdbc.core.RowMapper;
 import com.techcourse.domain.User;
-import java.sql.PreparedStatement;
-import java.sql.ResultSet;
-import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.List;
 import javax.sql.DataSource;
@@ -25,162 +25,116 @@ public class UserDao {
     }
 
     public void insert(final User user) {
+        final String sql = "insert into users (account, password, email) values (?, ?, ?)";
+        final PreparedStatementSetter pss = pstmt -> {
+            pstmt.setString(1, user.getAccount());
+            pstmt.setString(2, user.getPassword());
+            pstmt.setString(3, user.getEmail());
+        };
         final JdbcTemplate insertJdbcTemplate = new JdbcTemplate() {
-            @Override
-            protected String createQuery() {
-                return "insert into users (account, password, email) values (?, ?, ?)";
-            }
-
             @Override
             protected DataSource getDataSource() {
                 return UserDao.this.getDataSource();
             }
-
-            @Override
-            protected void setValues(final PreparedStatement pstmt) throws SQLException {
-                pstmt.setString(1, user.getAccount());
-                pstmt.setString(2, user.getPassword());
-                pstmt.setString(3, user.getEmail());
-            }
-
-            @Override
-            protected Object mapRow(final ResultSet rs) throws SQLException {
-                return null;
-            }
         };
 
-        insertJdbcTemplate.update();
+        insertJdbcTemplate.update(sql, pss);
     }
 
     public void update(final User user) {
+        final String sql = "update users set account = ?, password = ?, email = ? where id = ?";
+        final PreparedStatementSetter pss = pstmt -> {
+            pstmt.setString(1, user.getAccount());
+            pstmt.setString(2, user.getPassword());
+            pstmt.setString(3, user.getEmail());
+            pstmt.setLong(4, user.getId());
+        };
         final JdbcTemplate updateJdbcTemplate = new JdbcTemplate() {
-            @Override
-            protected String createQuery() {
-                return "update users set account = ?, password = ?, email = ? where id = ?";
-            }
-
             @Override
             protected DataSource getDataSource() {
                 return UserDao.this.getDataSource();
             }
-
-            @Override
-            protected void setValues(final PreparedStatement pstmt) throws SQLException {
-                pstmt.setString(1, user.getAccount());
-                pstmt.setString(2, user.getPassword());
-                pstmt.setString(3, user.getEmail());
-                pstmt.setLong(4, user.getId());
-            }
-
-            @Override
-            protected Object mapRow(final ResultSet rs) throws SQLException {
-                return null;
-            }
         };
 
-        updateJdbcTemplate.update();
+        updateJdbcTemplate.update(sql, pss);
     }
 
     public List<User> findAll() {
-        final JdbcTemplate selectJdbcTemplate = new JdbcTemplate() {
-            @Override
-            protected String createQuery() {
-                return "select id, account, password, email from users";
+        final String sql = "select id, account, password, email from users";
+        final RowMapper rowMapper = rs -> {
+            final List<User> users = new ArrayList<>();
+            while (rs.next()) {
+                final User user = new User(
+                        rs.getLong(1),
+                        rs.getString(2),
+                        rs.getString(3),
+                        rs.getString(4)
+                );
+                users.add(user);
             }
-
+            return users;
+        };
+        final PreparedStatementSetter pss = pstmt -> {
+        };
+        final JdbcTemplate selectJdbcTemplate = new JdbcTemplate() {
             @Override
             protected DataSource getDataSource() {
                 return UserDao.this.getDataSource();
             }
-
-            @Override
-            protected Object mapRow(final ResultSet rs) throws SQLException {
-                final List<User> users = new ArrayList<>();
-                while (rs.next()) {
-                    final User user = new User(
-                            rs.getLong(1),
-                            rs.getString(2),
-                            rs.getString(3),
-                            rs.getString(4)
-                    );
-                    users.add(user);
-                }
-                return users;
-            }
-
-            @Override
-            protected void setValues(final PreparedStatement pstmt) throws SQLException {
-            }
         };
 
-        return (List<User>) selectJdbcTemplate.query();
+        return (List<User>) selectJdbcTemplate.query(sql, pss, rowMapper);
     }
 
     public User findById(final Long id) {
-        final JdbcTemplate selectJdbcTemplate = new JdbcTemplate() {
-            @Override
-            protected String createQuery() {
-                return "select id, account, password, email from users where id = ?";
+        final String sql = "select id, account, password, email from users where id = ?";
+        final RowMapper rowMapper = rs -> {
+            if (rs.next()) {
+                return new User(
+                        rs.getLong(1),
+                        rs.getString(2),
+                        rs.getString(3),
+                        rs.getString(4)
+                );
             }
-
+            return null;
+        };
+        final PreparedStatementSetter pss = pstmt -> {
+            pstmt.setLong(1, id);
+        };
+        final JdbcTemplate selectJdbcTemplate = new JdbcTemplate() {
             @Override
             protected DataSource getDataSource() {
                 return UserDao.this.getDataSource();
             }
-
-            @Override
-            protected Object mapRow(final ResultSet rs) throws SQLException {
-                if (rs.next()) {
-                    return new User(
-                            rs.getLong(1),
-                            rs.getString(2),
-                            rs.getString(3),
-                            rs.getString(4)
-                    );
-                }
-                return null;
-            }
-
-            @Override
-            protected void setValues(final PreparedStatement pstmt) throws SQLException {
-                pstmt.setLong(1, id);
-            }
         };
 
-        return (User) selectJdbcTemplate.query();
+        return (User) selectJdbcTemplate.query(sql, pss, rowMapper);
     }
 
     public User findByAccount(final String account) {
-        final JdbcTemplate selectJdbcTemplate = new JdbcTemplate() {
-            @Override
-            protected String createQuery() {
-                return "select id, account, password, email from users where account = ?";
+        final String sql = "select id, account, password, email from users where account = ?";
+        final RowMapper rowMapper = rs -> {
+            if (rs.next()) {
+                return new User(
+                        rs.getLong(1),
+                        rs.getString(2),
+                        rs.getString(3),
+                        rs.getString(4)
+                );
             }
-
+            return null;
+        };
+        final PreparedStatementSetter pss = pstmt -> {
+            pstmt.setString(1, account);
+        };
+        final JdbcTemplate selectJdbcTemplate = new JdbcTemplate() {
             @Override
             protected DataSource getDataSource() {
                 return UserDao.this.getDataSource();
             }
-
-            @Override
-            protected Object mapRow(final ResultSet rs) throws SQLException {
-                if (rs.next()) {
-                    return new User(
-                            rs.getLong(1),
-                            rs.getString(2),
-                            rs.getString(3),
-                            rs.getString(4)
-                    );
-                }
-                return null;
-            }
-
-            @Override
-            protected void setValues(final PreparedStatement pstmt) throws SQLException {
-                pstmt.setString(1, account);
-            }
         };
 
-        return (User) selectJdbcTemplate.query();
+        return (User) selectJdbcTemplate.query(sql, pss, rowMapper);
     }
 }
