+++ jdbc/src/main/java/com/interface21/jdbc/core/PreparedStatementSetter.java
@@ -1,0 +1,9 @@
+package com.interface21.jdbc.core;
+
+import java.sql.PreparedStatement;
+import java.sql.SQLException;
+
+public interface PreparedStatementSetter {
+
+    void setValues(PreparedStatement pstmt) throws SQLException;
+}
