+++ jdbc/src/main/java/com/interface21/jdbc/core/RowMapper.java
@@ -2,11 +2,8 @@ package com.interface21.jdbc.core;
 
 import java.sql.ResultSet;
 import java.sql.SQLException;
-import javax.annotation.Nullable;
 
-@FunctionalInterface
-public interface RowMapper<T> {
+public interface RowMapper {
 
-    @Nullable
-    T mapRow(ResultSet rs, int rowNum) throws SQLException;
+    Object mapRow(ResultSet resultSet) throws SQLException;
 }
